+++ cli.js
@@ -12,6 +12,7 @@ const argOpts = {
 		help: ['h'],
 		analyse: ['a', 'analyze'],
 		vendored: ['v', 'vendor'],
+		gitattributes: ['g'],
 	},
 	boolean: ['full', 'help', 'analyze', 'vendored'],
 }
@@ -19,16 +20,17 @@ const args = yargs(process.argv.slice(2), argOpts);
 
 if (args.help) {
 	console.log(`\n${indent(1)}linguist usage:`);
-	usage(`linguist --analyse [<folder>] [--full] [--vendored]`, [
+	usage(`linguist --analyse [<folder>] [--full] [--vendored] [--gitattributes]`, [
 		`Analyse the language of all files found in a folder.`,
-		`<folder>   The folder to analyse (optional; default './')`,
-		`--full     List every file parsed (optional)`,
-		`--vendored Prevent skipping over vendored files (optional)`,
+		`<folder>   		The folder to analyse (optional; default './')`,
+		`--full     		List every file parsed (optional)`,
+		`--vendored 		Prevent skipping over vendored files (optional)`,
+		`--gitattributes 	Check .gitattributes files for custom file associations (optional)`,
 	].join('\n'));
 }
 else if (args.analyse) {
 	(async () => {
-		const data = await linguist(args._[0], { keepVendored: args.vendored });
+		const data = await linguist(args._[0], { keepVendored: args.vendored, checkAttributes: args.gitattributes });
 		let loggedData = data;
 		if (!args.full) delete loggedData.results;
 		console.log(loggedData);